ejercicio 10962 arreglado

+++ EjerciciosOmegaUp/Pares.java
@@ -5,13 +5,12 @@ import java.util.Scanner;
 public class Pares {
 
     public static void main(String[] args){
-        //FIXME
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int values[] = new int[n];
+        Long values[] = new Long[n];
         boolean flag = false;
         for (int i = 0; i < values.length; i++) {
-            values[i] = sc.nextInt();
+            values[i] = sc.nextLong();
             if (values[i] % 2 == 0) {
                 flag = true;
             }
